src/snitch/requirements.txt
anyio
asyncclick
tox
coverage
aiohttp
pyyaml